# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API",
        "description": "The Pet Shop API offers the required HTTP request methods and endpoints to meet the demands of a FE team creating a suitable UI",
        "contact": {
            "email": "anthonyjoboy2016@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/v1/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Authenticate administrative user",
                "description": "Returns the authorization information for an authenticated administrative user.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "required": [
                                            "email",
                                            "password"
                                        ],
                                        "properties": {
                                            "email": {
                                                "description": "Administrator E-mail",
                                                "type": "string"
                                            },
                                            "password": {
                                                "description": "Administrator Password",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "admin@buckhill.co.uk",
                                    "password": "********"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login was successful."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0Ojk3MDAvYXBpL3YxL2FkbWluL2xvZ2luIiwiaWF0IjoxNjYwMDkxNTk0LCJleHAiOjE2NjAwOTUxOTQsIm5iZiI6MTY2MDA5MTU5NCwianRpIjoiR3hHOVJWWjExV0syaGtJUCIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3IiwidXNlcl91dWlkIjoiZGU4MWZmOWItNTQ4MC00OGEyLTlmNTQtZGE3OTU3ZTQwMTFmIn0.i5ilpi9xXkOi9TdnfcYNRUkxqbpagtqCKbmFVj1Xyg0c"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "firstName": {
                                            "type": "string",
                                            "example": "Chris"
                                        },
                                        "lastName": {
                                            "type": "string",
                                            "example": "Pine"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "admin@buckhill.co.uk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token generation failed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Route Not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/admin/logout": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Logout administrator",
                "description": "Logs out the current authenticated administrative user.",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully logged out."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/admin/refresh": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Refresh token",
                "description": "Returns new token information.",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login was successful."
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0Ojk3MDAvYXBpL3YxL2FkbWluL2xvZ2luIiwiaWF0IjoxNjYwMDkxNTk0LCJleHAiOjE2NjAwOTUxOTQsIm5iZiI6MTY2MDA5MTU5NCwianRpIjoiR3hHOVJWWjExV0syaGtJUCIsInN1YiI6IjEiLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3IiwidXNlcl91dWlkIjoiZGU4MWZmOWItNTQ4MC00OGEyLTlmNTQtZGE3OTU3ZTQwMTFmIn0.i5ilpi9xXkOi9TdnfcYNRUkxqbpagtqCKbmFVj1Xyg0c"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "firstName": {
                                            "type": "string",
                                            "example": "Chris"
                                        },
                                        "lastName": {
                                            "type": "string",
                                            "example": "Pine"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "admin@buckhill.co.uk"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}